spawn `check_and_delete_old_databases`

+++ zebra-state/src/config.rs
@@ -132,16 +132,20 @@ impl Default for Config {
 /// - The directory name has a prefix `v`.
 /// - The directory name without the prefix can be parsed as an unsigned number.
 /// - The parsed number is lower than the hardcoded `DATABASE_FORMAT_VERSION`.
-pub fn check_and_delete_old_databases(cache_dir: PathBuf) {
-    let cache_dir = cache_dir.join("state");
-    if let Some(read_dir) = read_dir(cache_dir.clone()) {
-        for entry in read_dir.flatten() {
-            if let Some(dir_name) = parse_dir_name(entry) {
-                if let Some(version_number) = parse_version_number(dir_name.clone()) {
-                    if version_number < crate::constants::DATABASE_FORMAT_VERSION {
-                        let delete_path = cache_dir.join(dir_name);
-                        if remove_dir_all(delete_path.clone()).is_ok() {
-                            info!("deleted outdated state directory {:?}", delete_path);
+pub async fn check_and_delete_old_databases(config: Config) {
+    if !config.ephemeral && config.delete_old_database {
+        info!("checking old database versions");
+
+        let cache_dir = config.cache_dir.join("state");
+        if let Some(read_dir) = read_dir(cache_dir.clone()) {
+            for entry in read_dir.flatten() {
+                if let Some(dir_name) = parse_dir_name(entry) {
+                    if let Some(version_number) = parse_version_number(dir_name.clone()) {
+                        if version_number < crate::constants::DATABASE_FORMAT_VERSION {
+                            let delete_path = cache_dir.join(dir_name);
+                            if remove_dir_all(delete_path.clone()).is_ok() {
+                                info!("deleted outdated state directory {:?}", delete_path);
+                            }
                         }
                     }
                 }
